requirements.txt
Django>=5.1
django-allauth[socialaccount,mfa]>=65.4.0
django-appconf
django-bootstrap4>=23.4
django-countries>=7.6
django-debug-toolbar>=4.4
django-extensions>=3.2
django-htmlmin
django-simple-history
# mysqlclient
django-taggit>=6.0
djangorestframework>=3.15
docutils
gunicorn
python-decouple
pylibravatar
pyyaml
whitenoise>=6.7
